requirements.txt
Flask==2.3.3
pandas==2.2.3
gunicorn==21.2.0
openpyxl==3.1.2
mysql-connector-python==8.2.0
unidecode==1.3.7
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.0
